Commit of the version 0.06 of Girlfriend Advance which was still in a Work In Progress status. It seems that only very minor changes had been done.

+++ gfa-core/src/main/java/gfa/dma/Dma1_2.java
@@ -12,10 +12,10 @@ public class Dma1_2
   }
 
   public void setSrcHRegister(short srcH)
-  {src = (src & 0x0000ffff) | ((int) (srcH & 0x0fff) << 16);}
+  {src = (src & 0x0000ffff) | ((srcH & 0x0fff) << 16);}
 
   public void setDstHRegister(short dstH)
-  {dst = (dst & 0x0000ffff) | ((int) (dstH & 0x07ff) << 16);}
+  {dst = (dst & 0x0000ffff) | ((dstH & 0x07ff) << 16);}
 
   public void setCountRegister(short count)
   {this.count = (short) (count & 0x3fff);}
